requirements.txt
Flask==3.0.0
Flask-SQLAlchemy
psycopg2-binary
google-cloud-storage
google-auth
PyJWT
Werkzeug
Flask-Cors==3.0.10
